reorganize test data and use go:embed to include it in the executable
