+++ testdata/data-files/server-side-eval/builtin-attrs-anonymous.yml
