+++ testdata/data-files/server-side-eval/builtin-attrs-general-strings.yml
