+++ testdata/package_info.go
@@ -1,0 +1,3 @@
+// Package testdata contains functions for reading test data files. The actual data files are in
+// subdirectories of testdata/data-files.
+package testdata
