+++ testdata/testmodel/package_info.go
@@ -1,0 +1,3 @@
+// Package testmodel contains schema types for data files used in parameterized tests. The
+// actual data files are in subdirectories of testdata/data-files.
+package testmodel
