--- testmodel/package_info.go
@@ -1,3 +1,0 @@
-// Package testmodel contains schema types for data files used in parameterized tests, and a
-// parser to read them.
-package testmodel
